+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.1] - 2022.08.27
+
+* Rename `BaseViewMode` to `ViewModel`
+* Model can implement `WidgetsBindingObserver`
+* Removed useless `observer` property
+
 ## [1.0.0] - 2022.08.20
 
 * Basic implementation